requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0
watchdog==4.0.2
rich==13.7.1
python-dotenv==1.0.1
